backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch==1.9.0
torchvision==0.10.0
opencv-python==4.5.3.56
numpy==1.21.2
networkx==2.6.3
segmentation-models-pytorch==0.2.1
Pillow==8.3.2 